Adding validations to Expense Model

+++ Finances.Application/Expense/ExpenseDto.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.ComponentModel.DataAnnotations;
 
 namespace Finances.Application.Expense
 {
     public class ExpenseDto
     {
-        public float Value { get; set; }
+        public decimal Value { get; set; }
         public string Category { get; set; } = default!;
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public string? Description { get; set; }
