+++ Finances.Domain/Entities/Expense.cs
@@ -10,7 +10,7 @@ namespace Finances.Domain.Entities
     public class Expense
     {
         public int Id { get; set; }
-        public float Value { get; set; }
+        public decimal Value { get; set; }
         public string Category { get; set; } = default!;
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public string? Description { get; set; }
